Move blob URL hook to separate file
